+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2019 Ian Butterworth, Yavuz Faruk Bakman
+Copyright (c) 2019 Ian Butterworth, Yavuz Faruk Bakman, Robert Luciani
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
